+++ autoload/omnipytent/vim_plugin.py
@@ -3,7 +3,7 @@ import vim
 import os
 
 from .tasks_file import TasksFile, get_tasks_file
-from .tasks import invoke_with_dependencies
+from .tasks import invoke_with_dependencies, prompt_and_invoke_with_dependencies
 from .async_execution import AsyncExecutor, AsyncCommand
 from .util import vim_eval, vim_repr
 
@@ -42,12 +42,10 @@ def _api_entry_point(command):
 
 
 def prompt_and_invoke():
-    # Not sure if this will ever get implemented. In Integrake invoking without
-    # command would have popped up the list of tasks and prompted the user to
-    # pick one. Since I implemented Vim command line completion for tasks I
-    # ended up never using it(it turned out to be quite a nuisance, in fact),
-    # so I see no reason to implement in in Omnipytent.
-    raise NotImplementedError()
+    tasks_file = get_tasks_file()
+    tasks_file.load_if_stale()
+    executor = AsyncExecutor(prompt_and_invoke_with_dependencies(tasks_file))
+    executor.run_next()
 
 
 def invoke(taskname, *args):