+++ CHANGELOG.md
@@ -17,10 +17,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Subtasks:
   - For options task - `<task>?` to print the choice and `<task>!` to clear it.
   - For windows task - `<task>.close` to close the window.
+- Bare `:OP` to open a selection UI for picking tasks.
 
 ### Fixed
 - Package structure in `omnipytent.ext`.
 - CtrlP selection UI with `fmt` and multiple entries that have the same `fmt`.
+- A weird `bytes`/`str` problem with `CHOOSE`
 
 ## 1.3.0 - 2018-11-16
 ### Added